+++ R/compile.R
@@ -1,11 +1,11 @@
 #' @title Compile a PRQL query into a SQL query
 #' @param prql_query,prql a PRQL query string.
-#' @param dialect a SQL dialect name to use. If not specified, the dialect contained in the query will be used.
+#' @param target a compile target name to use. If not specified, the target contained in the query will be used.
 #' @param format a logical flag. Whether to format the SQL query.
 #' @param signature_comment a logical flag. Whether to add a signature comment to the output SQL query.
 #' @return a SQL query string
 #' @details [prql_to_sql] is deprecated in favor of [prql_compile].
-#' @seealso [prql_available_dialects]
+#' @seealso [prql_get_targets]
 #' @examples
 #' "from mtcars | filter cyl > 6 | select [cyl, mpg]" |>
 #'   prql_compile()
@@ -18,10 +18,10 @@
 #' filter cyl > 6
 #' select ![cyl]
 #' " |>
-#'   prql_compile("duckdb") |>
+#'   prql_compile("sql.duckdb") |>
 #'   cat()
 #'
-#' # Target dialect can also be written in the PRQL query.
+#' # Target can also be written in the PRQL query.
 #' "
 #' prql target:sql.duckdb
 #'
@@ -32,40 +32,37 @@
 #'   prql_compile() |>
 #'   cat()
 #' @export
-prql_compile <- function(prql_query, dialect = NA, format = TRUE, signature_comment = TRUE) {
-  if (length(dialect) == 1 && !(dialect %in% c(NA, dialects))) {
-    stop("Unsupported dialect. Please check with the 'prql_available_dialects()' function for available dialects.")
+prql_compile <- function(
+    prql_query,
+    target = getOption("prqlr.target"),
+    format = getOption("prqlr.format", TRUE),
+    signature_comment = getOption("prqlr.signature_comment", TRUE)) {
+  if (length(target) == 1 && !(target %in% c(NA, prql_get_targets()))) {
+    paste0(
+      r"(Unsupported target `")",
+      target,
+      r"("`. Please check with the `prql_get_targets()` function for available targets.)"
+    ) |>
+      stop()
   }
-  compile(prql_query, dialect, format, signature_comment)
-}
 
-dialects <- c(
-  "ansi",
-  "bigquery",
-  "clickhouse",
-  "generic",
-  "hive",
-  "mssql",
-  "mysql",
-  "postgres",
-  "sqlite",
-  "snowflake",
-  "duckdb"
-)
-
-#' @title Available dialect names
-#' @description Available dialects for the `dialect` option of the [prql_compile()] function.
-#' @return a character vector of dialect names.
-#' @examples
-#' prql_available_dialects()
-#' @export
-prql_available_dialects <- function() {
-  dialects
+  compile(prql_query, target, format, signature_comment) |>
+    unwrap()
 }
 
 #' @rdname prql_compile
 #' @export
 prql_to_sql <- function(prql) {
   .Deprecated("prql_compile")
+
   prql_compile(prql)
+}
+
+#' @title prql-compiler's version
+#' @return a [numeric_version] with the version of the built-in prql-compiler.
+#' @examples
+#' prql_version()
+#' @export
+prql_version <- function() {
+  numeric_version(compiler_version())
 }