+++ tests/testthat/test-compile.R
@@ -1,3 +1,16 @@
+test_that("target set in the header ", {
+  mssql_query <- "prql target:sql.mssql\nfrom a | take 1"
+
+  expect_equal(
+    prql_compile(mssql_query, format = FALSE, signature_comment = FALSE),
+    "SELECT TOP (1) * FROM a"
+  )
+  expect_equal(
+    prql_compile(mssql_query, target = "sql.generic", format = FALSE, signature_comment = FALSE),
+    "SELECT * FROM a LIMIT 1"
+  )
+})
+
 test_that("Not a string object", {
   expect_error(
     1 |> prql_compile(),
@@ -9,11 +22,26 @@ test_that("Not a string object", {
   )
 })
 
-test_that("Unsupported dialect", {
+test_that("Unsupported target", {
   expect_error(
     prql_compile("from a | select [b]", "foo"),
-    r"(Please check with the 'prql_available_dialects\(\)' function)"
+    r"(Please check with the `prql_get_targets\(\)` function)"
   )
+  expect_error(
+    prql_compile("prql target:foo\nfrom a | select [b]"),
+    r"(target `"foo"` not found)"
+  )
+    expect_error(
+    prql_compile("prql target:sql.foo\nfrom a | select [b]"),
+    r"(target `"sql.foo"` not found)"
+  )
+})
+
+test_that("Options", {
+  withr::local_options(list(prqlr.target = "sql.mssql", prqlr.format = FALSE, prqlr.signature_comment = FALSE))
+  expect_equal(prql_compile("from a | take 1"), "SELECT TOP (1) * FROM a")
+  withr::local_options(list(prqlr.target = "sql.postgres", prqlr.format = FALSE, prqlr.signature_comment = FALSE))
+  expect_equal(prql_compile("from a | take 1"), "SELECT * FROM a LIMIT 1")
 })
 
 test_that("PRQL query", {
@@ -24,12 +52,13 @@ test_that("PRQL query", {
     select [star_wars.*]
     select ![jar_jar_binks, midichlorians]"
     |>
-      compile("duckdb", TRUE, TRUE) |>
+      compile("sql.duckdb", TRUE, TRUE) |>
+      unwrap() |>
       cat()
   )
 })
 
-patrick::with_parameters_test_that("Dialects",
+patrick::with_parameters_test_that("Targets",
   {
     query <- "
 from flights
@@ -44,7 +73,11 @@ group [origin, dest] (
 sort [-origin, avg_delay]
 take 2
 "
-    expect_snapshot(cat(prql_compile(query, dialect, TRUE, FALSE)))
+    expect_snapshot(cat(prql_compile(query, target, TRUE, FALSE)))
   },
-  dialect = prql_available_dialects()
+  target = prql_get_targets()
 )
+
+test_that("prql-compiler's version", {
+  expect_equal(prql_version(), numeric_version("0.5.1"))
+})